Rename: 컨트롤러 경로 path에 맞게 수정
