+++ src/main/java/com/knu/KnowcKKnowcK/controller/articleSummary/SummaryController.java
@@ -3,10 +3,15 @@ package com.knu.KnowcKKnowcK.controller.articleSummary;
 import com.knu.KnowcKKnowcK.apiResponse.ApiResponseDto;
 import com.knu.KnowcKKnowcK.apiResponse.SuccessCode;
 import com.knu.KnowcKKnowcK.dto.requestdto.SummaryRequestDto;
+import com.knu.KnowcKKnowcK.dto.responsedto.MySummaryResponseDto;
 import com.knu.KnowcKKnowcK.dto.responsedto.SummaryHistoryResponseDto;
 import com.knu.KnowcKKnowcK.dto.responsedto.SummaryResponseDto;
+import com.knu.KnowcKKnowcK.enums.Status;
+import com.knu.KnowcKKnowcK.service.myPage.MySummaryService;
 import com.knu.KnowcKKnowcK.service.summary.SummaryService;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.Parameters;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -15,6 +20,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+
 @RestController
 @RequiredArgsConstructor
 @RequestMapping(value = "/api/summary", produces = "application/json;charset=UTF-8")
@@ -22,6 +29,7 @@ import org.springframework.web.bind.annotation.*;
 public class SummaryController {
 
     private final SummaryService summaryService;
+    private final MySummaryService mySummaryService;
 
 
     @GetMapping("/load")
@@ -62,5 +70,21 @@ public class SummaryController {
 
         SummaryResponseDto summaryResponseDto = summaryService.getSummaryFeedback(dto, authentication.getName());
         return ApiResponseDto.success(SuccessCode.OK, summaryResponseDto);
+    }
+
+    @GetMapping("")
+    @Operation(summary="사용자 요약",description="사용자가 작성했거나 작성한 요약 조회")
+    @Parameters({
+            @Parameter(name = "status",description = "원하는 요약이 작성 중인지,작성 한건지",example = "DONE"),
+    })
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "작성한 요약 조회 성공"),
+            @ApiResponse(responseCode = "400", description = "작성한 요약 조회 실패")
+    })
+    public ApiResponseDto<List<MySummaryResponseDto>> getMySummary(
+            Authentication authentication,
+            @RequestParam(value = "status",defaultValue = "DONE") Status status
+    ){
+        return ApiResponseDto.success(SuccessCode.OK, mySummaryService.getMySummaries(authentication.getName(),status));
     }
 }
